+++ modules/cacher.js
@@ -4,19 +4,35 @@ const NodeCache = require( "node-cache" );
 const localDB = new NodeCache();
 const config = require('../config');
 const format = require('string-format');
+var https = require('https');
+var fs = require('fs');
+const mimetypes = require('mime-types');
 
-let _recaptcha_version = undefined;
-const RECAPTCHA_API = 'https://lastgrupdate.com/api/last-update?version_only=web_cacher';
+/**
+ * Downloda buffer
+ * @param url
+ * @returns {any | promise | Promise<any>}
+ */
+var download_buffer = function(url) {
+    let deferred = Q.defer();
+    var request = https.get(url, function(response) {
+        let mimetype = mimetypes.lookup(url.split('?')[0]);
+        deferred.resolve({content: response, mimetype: mimetype});
+    }).on('error', function(err) { // Handle errors
+        deferred.reject(new Error(err.message || err));
+    });
+    return deferred.promise;
+};
+
 
 /**
  * Download URL
  * @param url
  */
-function download_url(url){
+function download_text(url){
     let deferred = Q.defer();
     // make request to get content
     requestify.get(url).then(function (response) {
-        // Get the response body
         let body = response.getBody();
         deferred.resolve(body);
     }).catch(function (err) {
@@ -25,13 +41,16 @@ function download_url(url){
     return deferred.promise;
 }
 
+let _recaptcha_version = undefined;
+const RECAPTCHA_API = 'https://lastgrupdate.com/api/last-update?version_only=web_cacher';
+
 /**
  * Get URL from storage
  * @param url
  */
 function get_from_storage(url){
     let deferred = Q.defer();
-    localDB.get( url, function( err, value ){
+    localDB.get( url, function( err, value){
         if( !err ) return deferred.resolve(value);
         if(err == null && value === undefined) return deferred.resolve(undefined);      // not set
         deferred.reject(err.message || err);
@@ -62,10 +81,11 @@ function flush_storage(){
  * Check if recaptcha didn't changed it's version
  */
 setInterval(function (){
-    download_url(RECAPTCHA_API).then(function (v){
+    download_text(RECAPTCHA_API).then(function (v){
         v = v.version;
         // if versions differ, flush local db
         if(v !== _recaptcha_version){
+            console.log('[+] Flushed internal DB'.bold.green);
             _recaptcha_version = v;
             flush_storage();
         }
@@ -81,15 +101,16 @@ module.exports = {
         let from_db = false;
         // try to get it from storage 1st
         get_from_storage(url).then(function (c){
-           if(!c) return download_url(url);     // if not set inside DB, download
+           if(!c) return download_buffer(url);     // if not set inside DB, download
            else {
                from_db = true;     // so we know not to re-set it
                return c;
            }                       // if set, do not re-download
         }).then(function (c){
             deferred.resolve(c);   // return content
+
             // if we didn't got it from DB, set it now
-            if(!from_db) return set_to_storage(url, c);     // set to storage
+            if(!from_db) return set_to_storage(url, c);
         }).catch(function (err){
             if(err.code > 400 && err.code < 500) deferred.reject(new Error(format('{} response from server', err.code)));
             else if(err.code > 500 && err.code < 600) deferred.reject(new Error(format('{} response from server', err.code)));