+++ changelog.md
@@ -1,3 +1,7 @@
+27.03.2019
+---------
+- updated ctim caching, to interpret files in response based on mimetype (text, image, etc)
+
 16.03.2019
 ----------
 - allowing caching only for `google.com` and `gstatic.com`